requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
